wap-ghst strategy
