+++ src/strategies/aavegotchi-agip-37-wap-ghst/README.md
@@ -1,0 +1,9 @@
+# Aavegotchi AGIP 37 Strategy WAP GHST
+
+## Description
+
+This snapshot strategy enables voting power for staked and unstaked wapGHST.
+
+## References
+
+Aavegotchi AGIP 37: https://snapshot.org/#/aavegotchi.eth/proposal/0x9923aab6825158ec2503d88e3ee2f9c5fbb12000581d06343ac9829aa59b66a6
