+++ src/strategies/aavegotchi-agip-37-wap-ghst/examples.json
@@ -2,7 +2,7 @@
   {
     "name": "AGIP 37",
     "strategy": {
-      "name": "aavegotchi-agip-37",
+      "name": "aavegotchi-agip-37-wap-ghst",
       "params": {
         "ghstAddress": "0x385Eeac5cB85A38A9a07A70c73e0a3271CfB54A7",
         "gltrStakingAddress": "0x1fE64677Ab1397e20A1211AFae2758570fEa1B8c",
@@ -28,7 +28,7 @@
       }
     },
     "network": "137",
-    "addresses": ["0x26cf02F892B04aF4Cf350539CE2C77FCF79Ec172", "0x027Ffd3c119567e85998f4E6B9c3d83D5702660c"],
+    "addresses": ["0x26cf02F892B04aF4Cf350539CE2C77FCF79Ec172", "0x027Ffd3c119567e85998f4E6B9c3d83D5702660c", "0xDEA88c9FE09106b58cA7c026c82383c56eE1E041", "0x3a564B24EffA1Cb7Dc836AD094BaD28e69FCa371"],
     "snapshot": 33423178
   }
 ]
