--- src/strategies/aavegotchi-agip-37/README.md
@@ -1,13 +1,0 @@
-# Aavegotchi AGIP 37 Strategy
-
-## Description
-
-This snapshot strategy enables voting power for amGHST, wapGHST, and GLTR staking LP pairs including GHST-USDC, GHST-MATIC, and the Alchemica pairs.
-
-## Implementation
-
-
-
-## References
-
-Aavegotchi AGIP 37: https://snapshot.org/#/aavegotchi.eth/proposal/0x9923aab6825158ec2503d88e3ee2f9c5fbb12000581d06343ac9829aa59b66a6
